[ADD] New module sale_order_credit_limit_approval
