--- credit_limit_sale_order_approval/i18n/es.po
@@ -1,103 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* plsr_credit_limit_sale_order_approval
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 8.0\n"
-"Report-Msgid-Bugs-To: \n"
-"POT-Creation-Date: 2016-01-22 13:16+0000\n"
-"PO-Revision-Date: 2016-01-22 14:18+0100\n"
-"Last-Translator: <>\n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: 8bit\n"
-"Plural-Forms: \n"
-"X-Generator: Poedit 1.5.4\n"
-
-#. module: credit_limit_sale_order_approval
-#: view:sale.order:credit_limit_sale_order_approval.view_order_credit_limit_form
-msgid "Approve Sale"
-msgstr "Aprobar venta"
-
-#. module: credit_limit_sale_order_approval
-#: view:sale.order:credit_limit_sale_order_approval.view_order_credit_limit_form
-msgid "Confirm Sale"
-msgstr "Confirmar venta"
-
-#. module: credit_limit_sale_order_approval
-#: code:addons/credit_limit_sale_order_approval/models/sale.py:41
-#, python-format
-msgid "Contact credit limit exceeded."
-msgstr "El crédito concedido del contacto ha sido sobrepasado."
-
-#. module: credit_limit_sale_order_approval
-#: view:sale.order:credit_limit_sale_order_approval.view_order_credit_limit_form
-msgid "Deny order"
-msgstr "Denegar venta"
-
-#. module: credit_limit_sale_order_approval
-#: code:addons/credit_limit_sale_order_approval/models/sale.py:60
-#, python-format
-msgid ""
-"Gives the status of the quotation or sales order.               \n"
-"The exception status is automatically set when a cancel                "
-"operation occurs in the invoice validation (Invoice "
-"Exception)                or in the picking list process (Shipping "
-"Exception).               \n"
-"The 'Waiting Schedule' status is set when the invoice is                "
-"confirmed but waiting for the scheduler to run on the order                "
-"date."
-msgstr ""
-"Indica el estado del presupuesto o del pedido de venta.\n"
-"El estado de excepción se establece automáticamente cuando ocurre una "
-"operación de cancelación en la validación de la factura (excepción de "
-"factura) o en el proceso de albaranado (excepción de envío).\n"
-"El estado 'Esperando fecha planificada' se establece cuando la factura está "
-"confirmada, pero se espera a que el planificador se ejecute en la fecha de "
-"pedido establecida."
-
-#. module: credit_limit_sale_order_approval
-#: model:ir.model,name:credit_limit_sale_order_approval.model_res_partner
-msgid "Partner"
-msgstr "Contacto"
-
-#. module: credit_limit_sale_order_approval
-#: model:ir.model,name:credit_limit_sale_order_approval.model_sale_order
-msgid "Sales Order"
-msgstr "Pedido de venta"
-
-#. module: credit_limit_sale_order_approval
-#: field:res.partner,set_credit_limit:0
-msgid "Set credit limit"
-msgstr "Establecer crédito concedido"
-
-#. module: credit_limit_sale_order_approval
-#: view:sale.order:credit_limit_sale_order_approval.view_order_credit_limit_form
-msgid "Warning"
-msgstr "Atención"
-
-#. module: credit_limit_sale_order_approval
-#: view:sale.order:credit_limit_sale_order_approval.view_order_credit_limit_form
-msgid "base.group_sale_manager"
-msgstr ""
-
-#. module: credit_limit_sale_order_approval
-#: view:sale.order:credit_limit_sale_order_approval.view_order_credit_limit_form
-msgid "oe_highlight"
-msgstr ""
-
-#. module: credit_limit_sale_order_approval
-#: view:sale.order:credit_limit_sale_order_approval.view_order_credit_limit_form
-#, fuzzy
-msgid "waiting_for_approval"
-msgstr "Esperando aprobación"
-
-#. module: credit_limit_sale_order_approval
-#: view:res.partner:credit_limit_sale_order_approval.view_set_credit_limit_res_partner_form
-msgid "{'invisible': [('set_credit_limit', '=', False)]}"
-msgstr ""
-
-#~ msgid "CREDIT LIMIT EXCEEDED"
-#~ msgstr "EL CREDITO CONCEDIDO DE ESTE CONTACTO HA SIDO SOBREPASADO"
