+++ sale_order_credit_limit_approval/i18n/es.po
@@ -1,0 +1,83 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* sale_order_credit_limit_approval
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 8.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2016-01-26 11:43+0000\n"
+"PO-Revision-Date: 2016-01-26 12:46+0100\n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: \n"
+"X-Generator: Poedit 1.5.4\n"
+
+#. module: sale_order_credit_limit_approval
+#: view:sale.order:sale_order_credit_limit_approval.view_order_credit_limit_form
+msgid "Approve Sale"
+msgstr "Aprobar venta"
+
+#. module: sale_order_credit_limit_approval
+#: field:res.partner,available_credit:0
+msgid "Available credit"
+msgstr "Crédito disponible"
+
+#. module: sale_order_credit_limit_approval
+#: view:sale.order:sale_order_credit_limit_approval.view_order_credit_limit_form
+msgid "Confirm Sale"
+msgstr "Confirmar venta"
+
+#. module: sale_order_credit_limit_approval
+#: code:addons/sale_order_credit_limit_approval/models/sale.py:40
+#, python-format
+msgid "Contact credit limit exceeded."
+msgstr "El límite del crédito concedido del contacto ha sido sobrepasado."
+
+#. module: sale_order_credit_limit_approval
+#: view:sale.order:sale_order_credit_limit_approval.view_order_credit_limit_form
+msgid "Deny order"
+msgstr "Denegar venta"
+
+#. module: sale_order_credit_limit_approval
+#: model:ir.model,name:sale_order_credit_limit_approval.model_res_partner
+msgid "Partner"
+msgstr "Contacto"
+
+#. module: sale_order_credit_limit_approval
+#: model:ir.model,name:sale_order_credit_limit_approval.model_sale_order
+msgid "Sales Order"
+msgstr "Pedido de venta"
+
+#. module: sale_order_credit_limit_approval
+#: field:res.partner,set_credit_limit:0
+msgid "Set credit limit"
+msgstr "Establecer crédito concedido"
+
+#. module: sale_order_credit_limit_approval
+#: view:sale.order:sale_order_credit_limit_approval.view_order_credit_limit_form
+msgid "Warning"
+msgstr "Aviso"
+
+#. module: sale_order_credit_limit_approval
+#: view:sale.order:sale_order_credit_limit_approval.view_order_credit_limit_form
+msgid "base.group_sale_manager"
+msgstr ""
+
+#. module: sale_order_credit_limit_approval
+#: view:sale.order:sale_order_credit_limit_approval.view_order_credit_limit_form
+msgid "oe_highlight"
+msgstr ""
+
+#. module: sale_order_credit_limit_approval
+#: view:sale.order:sale_order_credit_limit_approval.view_order_credit_limit_form
+msgid "waiting_for_approval"
+msgstr ""
+
+#. module: sale_order_credit_limit_approval
+#: view:res.partner:sale_order_credit_limit_approval.view_set_credit_limit_res_partner_form
+msgid "{'invisible': [('set_credit_limit', '=', False)]}"
+msgstr ""
