+++ sale_order_credit_limit_approval/static/description/icon.png
