+++ sale_order_credit_limit_approval/static/description/index.html
@@ -1,7 +1,7 @@
 <section class="oe_container">
     <div class="oe_row oe_spaced">
         <div class="oe_span12">
-            <h2 class="oe_slogan">Credit Limit Sale Order Approval</h2>
+            <h2 class="oe_slogan">Sale Order Credit Limit Approval</h2>
             <p>This module was written to throw a warning when the credit limit of a customer is exceeded in a sale order, in addition to moving in this case the status to a new one named 'Waiting for approval'. If the credit limit of a customer is exceeded, only the sales manager will be able to approve the sale order.</p>
             <p>The module also adds a new option to make optional the use of a credit limit, and shows the current state of this if it has been activated.</p>
         </div>
