+++ sale_order_credit_limit_approval/views/res_partner_view.xml
